+++ changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- `redstoneSide` config option can now be a table, if you wish to have multiple outputs for redstone heartbeat
+
 ## 1.0.1
 ### Fixed
 - Fixed broken error handler